Split cli.py, remove the UDM_* names, commit 0.9.0 to a package
Three changes, in one commit because they interlock through CHANGELOG.md,
CLAUDE.md, TODO.md and tests/test_render.py, and splitting them would
mean an intermediate commit with a red suite.

Split cli.py into two modules, not the one CLAUDE.md guessed at.
artwork.py resolves nodes to images and imports no renderer; output.py
writes files and imports every renderer. They share nothing, so "a
reason per file" is satisfied, which was the real criterion rather than
the module count. cli.py keeps argument parsing, credential resolution,
logging setup and the cmd_* functions, and goes 1367 to 1027 lines.
obtain_icon_font() also stopped taking an argparse.Namespace and now
takes the three settings it reads, which is the layering fix in
miniature. No behaviour change; the demo renders identically.

Removed the UDM_* environment names, warned since 0.7.0. A UDM_*-only
credential file now gets the ordinary missing-configuration error naming
the variable to set, which is right: by that point the old name is not
part of the interface. layout.py still strips UDM_API_KEY from
Graphviz's environment, deliberately, because not reading a variable
does nothing about somebody who still exports one. There is a test
pinning that, since removing every mention is the obvious tidy-up.

Committed 0.9.0 to producing an installable artifact (KAN-132). Moved in
TODO.md from "Undecided" to "Committed to a version", which by that
file's own rule means a version named in the code, the tests and the
changelog. It reopens the lock-file decline, whose stated reason was
that the benefit is dev-only.
