+++ scripts/generate_manpage.py
@@ -167,9 +167,9 @@ def render() -> str:
         "\\fBUNIFI_MAP_ENV\\fR",
         "Path to a credential file, tried before the default locations.",
         ".PP",
-        "The \\fBUDM_\\fR\\fI*\\fR spellings of these are deprecated and warn once per",
-        "run. A key is never written into the environment, and is removed from the",
-        "environment Graphviz is run with.",
+        "The \\fBUDM_\\fR\\fI*\\fR spellings of these were removed in 0.9.0. A key is",
+        "never written into the environment, and is removed from the environment",
+        "Graphviz is run with.",
         ".SH FILES",
         ".TP",
         "\\fB.env\\fR, \\fB~/.config/unifi\\-map/env\\fR",
