+++ unifi-map.1
@@ -191,9 +191,9 @@ Which site to read. Overridden by \fB\-\-site\fR. Defaults to
 \fBUNIFI_MAP_ENV\fR
 Path to a credential file, tried before the default locations.
 .PP
-The \fBUDM_\fR\fI*\fR spellings of these are deprecated and warn once per
-run. A key is never written into the environment, and is removed from the
-environment Graphviz is run with.
+The \fBUDM_\fR\fI*\fR spellings of these were removed in 0.9.0. A key is
+never written into the environment, and is removed from the environment
+Graphviz is run with.
 .SH FILES
 .TP
 \fB.env\fR, \fB~/.config/unifi\-map/env\fR
